requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
requests==2.32.3
python-dotenv==1.0.1
openai==1.63.2